Website Branding Fixes
Deltik - with apologies for the erroneous image. Not sure how that happened, have replaced it with the new one - do let me know if it complains again.

Have also corrected some line height errors in mbr-additional.css resulting in text rendering incorrectly on mobile. Hopefully it behaves itself this time!